+++ tests/CMakeLists.txt
@@ -310,6 +310,8 @@ set(ARANGODB_TESTS_SOURCES
   ${ADDITIONAL_TEST_SOURCES})
 
 set(ARANGODB_REPLICATION2_TEST_SOURCES
+  Replication2/Helper/ModelChecker/Actors.cpp
+  Replication2/Helper/ModelChecker/AgencyTransition.cpp
   Replication2/Helper/WaitForQueue.cpp
   Replication2/Mocks/AsyncFollower.cpp
   Replication2/Mocks/AsyncFollower.h
@@ -323,8 +325,13 @@ set(ARANGODB_REPLICATION2_TEST_SOURCES
   Replication2/Mocks/FakeReplicatedLog.h
   Replication2/Mocks/FakeReplicatedState.cpp
   Replication2/Mocks/PersistedLog.cpp
+  Replication2/Mocks/PersistedLog.cpp
+  Replication2/Mocks/PersistedLog.h
   Replication2/Mocks/PersistedLog.h
   Replication2/Mocks/ReplicatedLogMetricsMock.cpp
+  Replication2/ModelChecker/ActorModel.cpp
+  Replication2/ModelChecker/ModelChecker.cpp
+  Replication2/ModelChecker/Tests/ModelCheckerTest.cpp
   Replication2/ReplicatedLog/AppendEntriesBatchTest.cpp
   Replication2/ReplicatedLog/AppendEntriesErrorReasonTest.cpp
   Replication2/ReplicatedLog/CalcCommitIndexTest.cpp
@@ -360,8 +367,8 @@ set(ARANGODB_REPLICATION2_TEST_SOURCES
   Replication2/ReplicatedState/ReplicatedStateFeatureTest.cpp
   Replication2/ReplicatedState/ReplicatedStateTest.cpp
   Replication2/ReplicatedState/StateMachines/MyStateMachine.cpp
-  Replication2/Mocks/PersistedLog.cpp
-  Replication2/Mocks/PersistedLog.h
+  Replication2/ReplicatedState/SupervisionSimulationTest2.cpp
+  Replication2/ReplicatedState/SupervisionTest.cpp
   Replication2/Serialization/AgencyLogSpecificationTest.cpp
   Replication2/Serialization/LogCommonTest.cpp
   Replication2/Serialization/LogStatusTest.cpp
@@ -369,9 +376,7 @@ set(ARANGODB_REPLICATION2_TEST_SOURCES
   Replication2/Streams/MultiplexerConcurrencyTest.cpp
   Replication2/Streams/TestLogSpecification.cpp
   Replication2/Streams/TestLogSpecification.h
-  Replication2/ReplicatedState/SupervisionTest.cpp
-        Replication2/ReplicatedState/SupervisionSimulationTest2.cpp
-        Replication2/ModelChecker/ModelChecker.cpp Replication2/ModelChecker/ModelChecker.h Replication2/ModelChecker/Tests/ModelCheckerTest.cpp Replication2/ModelChecker/ActorModel.cpp Replication2/ModelChecker/ActorModel.h)
+)
 
 
 if (LINUX)
